test(vault-encryption): Sub-E (#43) ペガサス工程5指摘③ファイル分割 + cache_relocked検証 (Issue #43)
## ペガサス工程5指摘解消 (致命③ ファイル分割 + 銀時143e8eb取り込み)

### 1. tests/sub_e_v2_integration.rs (693行) を sub_e_v2_integration/ 配下に責務分割
旧単一 693 行ファイルを 7 ファイルに再編、各 250 行以内に収めた:
- helpers.rs (134) — 共通ヘルパ DRY 集約
- unlock.rs (48) — TC-E-I01
- backoff.rs (144) — TC-E-I02 / I02b
- lock_lifecycle.rs (103) — TC-E-I04 / I05
- handshake.rs (98) — TC-E-I07 / I09
- rekey_rotate.rs (245) — TC-E-I06 / I06b / I08
- sanity.rs (15) — TempDir lifecycle

entry (~50 行) は #[path = "..."] で各 sub module を明示パス指定、cargo は
単一 integration test binary としてビルド (Rust の慣習)。テスト名 prefix が
backoff:: / handshake:: / rekey_rotate:: 等で表示されて読みやすくなった。

### 2. 銀時 commit 143e8eb (cache_relocked: bool 追加) のテスト取り込み
- TC-E-I08 修正: Rekeyed { cache_relocked, .. } で cache_relocked == true assert
- TC-E-I06b 新設: RotateRecovery 正常経路で cache_relocked: true 確認 +
  cache.is_unlocked が維持されることを機械固定 (Lie-Then-Surprise 経路の
  構造防衛、ペガサス致命指摘①の確認テスト側担保)

### 3. test-design.md §14.13.3 ペガサス工程5指摘解消 SSoT 追記

## ローカル検証

- cargo fmt --check: PASS
- cargo clippy --all-targets --all-features: 0 errors
- 静的検査 TC-E-S01..S08: 8/8 PASS
- daemon lib: 64 PASS
- sub_e_v2_integration (11件 = 旧10件 + TC-E-I06b): 11/11 PASS
- core lib (185件、銀時 cache_relocked: false 検証 2件含む): 全 PASS
- infra lib + vault_migration_integration: 108 PASS
- 合計 376 件 全 PASS

ペガサス致命指摘①③ 解消。②は ux-and-msg.md 担当のセル別 commit。
